Start work on template for HTML reporting
A basic template has been added to the report-template directory,
which will contain a completely static example of a HTML report.
Currently the template includes a file browser, and shows basic
information about a particular run of cargo-tarpaulin.

The template still needs additional work before it shows everything
that is needed. Coverage information for directories and files is
not shown yet, the template is not as accomodating to smaller screens
as it could be, and some styling may be performed.